# Rewrite `if` operator with while loop